Move `assignDiscordRole` to `service-discord`
