+++ apps/gov-website/pages/api/identity/judgement.tsx
@@ -8,7 +8,7 @@ import {
 	isIdentityValueMatched,
 	signAndSend,
 } from "@app-gov/service/cennznet";
-import { getDiscordBot } from "@app-gov/service/discord";
+import { assignDiscordRole } from "@app-gov/service/discord";
 import {
 	CENNZ_NETWORK,
 	DISCORD_WEBSITE_BOT,
@@ -70,7 +70,13 @@ export default withMethodGuard(
 			]);
 
 			// 4. Assign user with a special role
-			if (DISCORD_WEBSITE_BOT.Token) await assignDiscordRole(discordUsername);
+			if (DISCORD_WEBSITE_BOT.Token)
+				await assignDiscordRole(
+					discordUsername,
+					DISCORD_WEBSITE_BOT.Token,
+					DISCORD_WEBSITE_BOT.ServerId,
+					DISCORD_WEBSITE_BOT.IdentityRoleId
+				);
 
 			return res.json({ ok: true });
 		} catch (error) {
@@ -84,35 +90,3 @@ export default withMethodGuard(
 	},
 	["POST"]
 );
-
-const assignDiscordRole = async (discordUsername: string) => {
-	const [username, discriminator] = discordUsername.split("#");
-
-	const discordBot = await getDiscordBot(DISCORD_WEBSITE_BOT.Token);
-
-	const guildCache = discordBot.guilds.cache.get(DISCORD_WEBSITE_BOT.ServerId);
-	if (!guildCache) throw { message: "DISCORD_SERVER_NOT_FOUND" };
-	await guildCache.members.fetch();
-
-	const user = guildCache.members.cache.find((user) => {
-		return (
-			user.user.username === username &&
-			user.user.discriminator === discriminator
-		);
-	});
-	if (!user) throw { message: "DISCORD_USER_NOT_FOUND" };
-
-	const identityRole = guildCache.roles.cache.find(
-		(role) => role.id === DISCORD_WEBSITE_BOT.IdentityRoleId
-	);
-	if (!identityRole) throw { message: "DISCORD_IDENTITY_ROLE_NOT_FOUND" };
-
-	await user.roles.add(identityRole);
-	// Send a message to the user letting them know the verification has been successful
-	await user.send(
-		`***Welcome Citizen ${username}.*** \n\n` +
-			`Thank you for doing your part in governing the CENNZnet blockchain!\n` +
-			`You have been assigned the ${identityRole.name} role and can now participate in private channels.\n` +
-			`Please note that for your safety, we will never ask for private keys, seed phrases or send links via DM.`
-	);
-};
