+++ libs/service/discord/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./assignDiscordRole";
 export * from "./getDiscordBot";
 export * from "./getDiscordUpdate";
 export * from "./getProposalFields";
